Refactor DremioFlightProducer tests to be able to run query tests with both statement and preparedstatement
