+++ services/arrow-flight/src/test/java/com/dremio/service/flight/AbstractTestFlightSqlServer.java
@@ -16,7 +16,11 @@
 
 package com.dremio.service.flight;
 
+import static java.util.Arrays.asList;
+
 import java.sql.SQLException;
+import java.util.Collection;
+import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.regex.Pattern;
 import java.util.Collections;
@@ -30,6 +34,8 @@ import org.apache.arrow.vector.VarCharVector;
 import org.apache.arrow.vector.VectorSchemaRoot;
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 
 import com.dremio.service.catalog.TableType;
 import com.dremio.exec.expr.fn.impl.RegexpUtil;
@@ -38,190 +44,43 @@ import com.dremio.exec.proto.UserProtos;
 
 import com.google.common.collect.ImmutableList;
 
+@RunWith(Parameterized.class)
 public abstract class AbstractTestFlightSqlServer extends AbstractTestFlightServer {
 
+  private final ExecutionMode executionMode;
+
+  enum ExecutionMode {
+    STATEMENT,
+    PREPARED_STATEMENT
+  }
+
+  @Parameterized.Parameters(name = "ExecutionMode: {0}")
+  public static Collection<ExecutionMode> parameters() {
+    return asList(ExecutionMode.STATEMENT, ExecutionMode.PREPARED_STATEMENT);
+  }
+
+  public AbstractTestFlightSqlServer(ExecutionMode executionMode) {
+    this.executionMode = executionMode;
+  }
+
+  private FlightInfo executeStatement(String query) {
+    FlightSqlClient client = getFlightClientWrapper().getSqlClient();
+    return client.execute(query, getCallOptions());
+  }
+
+  private FlightInfo executePreparedStatement(String query) throws SQLException {
+    FlightSqlClient client = getFlightClientWrapper().getSqlClient();
+    FlightSqlClient.PreparedStatement preparedStatement = client.prepare(query, getCallOptions());
+    return preparedStatement.execute(getCallOptions());
+  }
+
   @Override
-  public FlightInfo getFlightInfo(String query) throws SQLException {
-    final FlightClientUtils.FlightClientWrapper clientWrapper = getFlightClientWrapper();
-    return clientWrapper.getSqlClient().execute(query, getCallOptions());
-  }
-
-  @Test
-  public void testGetCatalogs() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    CallOption[] callOptions = getCallOptions();
-
-    FlightInfo flightInfo = flightSqlClient.getCatalogs(callOptions);
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(), callOptions)) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-      Assert.assertEquals(1, root.getRowCount());
-
-      String catalogName = ((VarCharVector) root.getVector("catalog_name")).getObject(0).toString();
-      Assert.assertEquals("DREMIO", catalogName);
+  public FlightInfo getFlightInfo(String query) throws SQLException{
+    switch (executionMode) {
+      case STATEMENT: return executeStatement(query);
+      case PREPARED_STATEMENT: return executePreparedStatement(query);
     }
-  }
 
-  @Test
-  public void testGetTablesWithoutFiltering() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getTables(null, null, null,
-      null, false, getCallOptions());
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(), getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      Assert.assertEquals(root.getRowCount(), 28);
-    }
-  }
-
-  @Test
-  public void testGetTablesFilteringByCatalogPattern() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getTables("DREMIO", null, null,
-      null, false, getCallOptions());
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(),
-      getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      Assert.assertEquals(root.getRowCount(), 28);
-    }
-  }
-
-  @Test
-  public void testGetTablesFilteringBySchemaPattern() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getTables(null, "INFORMATION_SCHEMA",
-      null, null, false, getCallOptions());
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(),
-      getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      Assert.assertEquals(root.getRowCount(), 5);
-    }
-  }
-
-  @Test
-  public void testGetTablesFilteringByTablePattern() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getTables(null, null, "COLUMNS",
-      null, false, getCallOptions());
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(),
-      getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      Assert.assertEquals(root.getRowCount(), 1);
-    }
-  }
-
-  @Test
-  public void testGetTablesFilteringByTableTypePattern() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getTables(null, null, null,
-      Collections.singletonList("SYSTEM_TABLE"), false, getCallOptions());
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(),
-      getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      Assert.assertEquals(root.getRowCount(), 28);
-    }
-  }
-
-  @Test
-  public void testGetTablesFilteringByMultiTableTypes() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getTables(null, null, null,
-      ImmutableList.of("TABLE", "VIEW"), false, getCallOptions());
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(),
-      getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      Assert.assertEquals(root.getRowCount(), 0);
-    }
-  }
-
-  @Test
-  public void testGetTablesTypes() throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getTableTypes(getCallOptions());
-    try (FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(), getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      final ImmutableList<TableType> tableTypes = ImmutableList.of(TableType.TABLE, TableType.SYSTEM_TABLE, TableType.VIEW);
-      final int counter = tableTypes.size();
-
-      final IntStream range = IntStream.range(0, counter);
-
-      Assert.assertEquals(root.getRowCount(), counter);
-      range.forEach(i -> {
-        Assert.assertEquals(root.getVector(0).getObject(i).toString(), tableTypes.get(i).toString());
-      });
-    }
-  }
-
-  private void testGetSchemas(String catalog, String schemaPattern, boolean expectNonEmptyResult) throws Exception {
-    FlightSqlClient flightSqlClient = getFlightClientWrapper().getSqlClient();
-    FlightInfo flightInfo = flightSqlClient.getSchemas(catalog, schemaPattern, getCallOptions());
-    try (
-      FlightStream stream = flightSqlClient.getStream(flightInfo.getEndpoints().get(0).getTicket(), getCallOptions())) {
-      Assert.assertTrue(stream.next());
-      VectorSchemaRoot root = stream.getRoot();
-
-      Predicate<String> catalogNamePredicate = MetadataProviderConditions.getCatalogNamePredicate(
-        catalog != null ? UserProtos.LikeFilter.newBuilder().setPattern(catalog).build() : null);
-      Pattern schemaFilterPattern = schemaPattern != null ? Pattern.compile(RegexpUtil.sqlToRegexLike(schemaPattern)) :
-        Pattern.compile(".*");
-
-      VarCharVector catalogNameVector = (VarCharVector) root.getVector("catalog_name");
-      VarCharVector schemaNameVector = (VarCharVector) root.getVector("schema_name");
-
-      for (int i = 0; i < root.getRowCount(); i++) {
-        String catalogName = catalogNameVector.getObject(i).toString();
-        String schemaName = schemaNameVector.getObject(i).toString();
-
-        Assert.assertTrue(catalogNamePredicate.test(catalogName));
-        Assert.assertTrue(schemaFilterPattern.matcher(schemaName).matches());
-      }
-
-      if (expectNonEmptyResult) {
-        Assert.assertTrue(root.getRowCount() > 0);
-      }
-    }
-  }
-
-  @Test
-  public void testGetSchemasWithNoFilter() throws Exception {
-    testGetSchemas(null, null, true);
-  }
-
-  @Test
-  public void testGetSchemasWithBothFilters() throws Exception {
-    testGetSchemas("DREMIO", "INFORMATION_SCHEMA", true);
-  }
-
-  @Test
-  public void testGetSchemasWithCatalog() throws Exception {
-    testGetSchemas("DREMIO", null, true);
-  }
-
-  @Test
-  public void testGetSchemasWithSchemaFilterPattern() throws Exception {
-    testGetSchemas(null, "sys", true);
-  }
-
-  @Test
-  public void testGetSchemasWithNonMatchingSchemaFilter() throws Exception {
-    testGetSchemas(null, "NON_EXISTING_SCHEMA", false);
-  }
-
-  @Test
-  public void testGetSchemasWithNonMatchingCatalog() throws Exception {
-    testGetSchemas("NON_EXISTING_CATALOG", null, false);
+    throw new IllegalStateException();
   }
 }
